Detection and Sync improvements
+ Added header with debug macros
+ Added warning for SF 6: the samples are actually wrapped around to SF
12
~ Unwrapped some function calls to liquid-dsp::liquid_unwrap_phase
~ Simplified some more code
~ Rewrote sliding_norm_cross_correlate_upchirp: Now we detect a falling
edge
and do a correlation from there to test the sync rate.
Also added code that can perform a perfect overlay shift to the ideal
chirp
(but this leaves the falling edge inside the window)
~ Rewrote find_preamble_start_fast: Now detects the start of the
preamble
by looking for for absolute complex values above a threshold
~ Fixed ids in qa_BasicTest_Data.xml
~ qa_BasicTest.log contains test runs showing the improvements in
passing tests.
+ Added createWhiteningValues.py to generate a sample with zeros and
append
the whitening sequence to a file so it can be averaged (untested).
This uses lora-whitening-sequencer.
~ Added zero padding to whitening sequences in tables.h